Allow ./signal-lamp dark
Toggling might be a really bad idea. We're not in change of what really
happens in any application after we've tried to "set a theme" for any
application and storing that theme name in a file.

An alternative route would be to just provide a list of available
themes, together with "this is the currently chosen theme" and "this is
the theme I'd auto-toggle to, if you don't want to build that logic
yourself, or provide a menu or such".